Variable width2 used at wrong place. Image size too small by one when passed to ResizerWindow. Use fractional arithmetic for aspect ratio calculations to avoid rounding errors.
git-svn-id: file:///srv/svn/repos/haiku/haiku/trunk@19542 a95241bf-73f2-0310-859d-f6bbb57e9c96

+++ src/apps/showimage/ShowImageWindow.h
@@ -78,8 +78,8 @@ class ShowImageWindow : public BWindow {
 		void PrepareForPrint();
 		void Print(BMessage *msg);
 		
-		void OpenResizerWindow(float width, float height);
-		void UpdateResizerWindow(float width, float height);
+		void OpenResizerWindow(int32 width, int32 height);
+		void UpdateResizerWindow(int32 width, int32 height);
 		void CloseResizerWindow();
 
 		BFilePanel *fSavePanel;
